+++ src/Recommend/Recommend.php
@@ -4,10 +4,9 @@ namespace srag\Plugins\SrTile\Recommend;
 
 use ilSrTilePlugin;
 use srag\DIC\SrTile\DICTrait;
-use srag\Notifications4Plugin\Notifications4Plugins\Notification\Repository as NotificationRepository;
-use srag\Notifications4Plugin\Notifications4Plugins\Sender\Repository as SenderRepository;
-use srag\Plugins\Notifications4Plugins\Notification\Language\NotificationLanguage;
-use srag\Plugins\Notifications4Plugins\Notification\Notification;
+use srag\Notifications4Plugin\SrTile\Utils\Notifications4PluginTrait;
+use srag\Plugins\SrTile\Notification\Notification\Language\NotificationLanguage;
+use srag\Plugins\SrTile\Notification\Notification\Notification;
 use srag\Plugins\SrTile\Tile\Tile;
 use srag\Plugins\SrTile\Utils\SrTileTrait;
 use Throwable;
@@ -23,6 +22,7 @@ class Recommend {
 
 	use DICTrait;
 	use SrTileTrait;
+	use Notifications4PluginTrait;
 	const PLUGIN_CLASS_NAME = ilSrTilePlugin::class;
 	/**
 	 * @var Tile
@@ -53,10 +53,9 @@ class Recommend {
 		try {
 			$mail_template = $this->tile->getRecommendMailTemplate();
 
-			$notification = NotificationRepository::getInstance(Notification::class, NotificationLanguage::class)
-				->getNotificationByName($mail_template);
+			$notification = self::notification(Notification::class, NotificationLanguage::class)->getNotificationByName($mail_template);
 
-			$sender = SenderRepository::getInstance()->factory()->externalMail("", $this->getRecommendedTo());
+			$sender = self::sender()->factory()->externalMail("", $this->getRecommendedTo());
 
 			$placeholders = [
 				"link" => $this->getLink(),
@@ -65,7 +64,7 @@ class Recommend {
 				"user" => self::dic()->user()
 			];
 
-			SenderRepository::getInstance()->send($sender, $notification, $placeholders, $placeholders["user"]->getLanguage());
+			self::sender()->send($sender, $notification, $placeholders, $placeholders["user"]->getLanguage());
 
 			return true;
 		} catch (Throwable $ex) {